// src/Models/Suggestions.ts
export interface ISuggested {
  address: string | null;
  avatar: string | null;
  birthDate: string | null;
  cityId: number | null;
  countryId: number | null;
  email: string;
  firstName: string;
  lastName: string;
  password: string;
  phoneNumber: number;
  profileText: string | null;
  userId: number;
  userProfile: string | null;
}